Fix unmocked requests
